Handle instructions :path field
A future PR will allow the testing of a subset of the response data.

In anticipation of that PR, this commit provides a way to target that subset of data. The Instructions class now returns a path specified in the instructions file.

For example, the response data might contain multiple members within an array:
```
{ members: [{id: 1, name: 'Jim' }, { id: 2, name: 'Joe' }] }
```
An array can be specified in the :path field. To target Joe, the array would be:

```YAML
:path:
    - :members
    - :id: 2
```
That is, the path specifies that the test wants a hash containing the id of 2 within the members array.