Handle instructions :target field
A future PR will allow the testing of a subset of the response data.

In anticipation of that PR, this commit provides a way to target that subset of data.

For example, the response data might contain multiple members within an array:
```
{ members: [{id: 1, name: 'Jim' }, { id: 2, name: 'Joe' }] }
```

To target the first member, we can supply a key value pair:
```YAML
target: id: 1
```

or

```YAML
target: name: 'Jim'
```